Update load kext script for x86_64 and arm64e

+++ load_kext.sh
@@ -1,17 +1,22 @@
 set -e
 
-sudo nvram boot-args="debug=0x100 keepsyms=1 amfi_get_out_of_my_way=1 amfi_allow_any_signature=1 --arm64e_preview_abi"
+if [ "$1" == "arm64e" ]; then
+	sudo nvram boot-args="debug=0x100 keepsyms=1 amfi_get_out_of_my_way=1 amfi_allow_any_signature=1 --arm64e_preview_abi"
+elif [ "$1" == "x86_64" ]; then
+	sudo nvram boot-args="debug=0x100 keepsyms=1 amfi_get_out_of_my_way=1 amfi_allow_any_signature=1"
+else
+	echo Bad Architecture $1
 
-export CFLAGS="-target arm64e-apple-macos12.0"
-export LDFLAGS="-target arm64e-apple-macos12.0"
+	exit 1
+fi
 
-sudo rm -R build/com.YungRaj.MacRootKit.kext
+if [ -d build/com.YungRaj.MacRootKit.kext ]; then
+	sudo rm -R build/com.YungRaj.MacRootKit.kext
+fi
 
 sudo /usr/bin/kmutil install --volume-root /
 
-make -f make_kext.mk clean
-
-make -f make_kext.mk
+./build_kext.sh $1
 
 sudo chmod -R 755 build/com.YungRaj.MacRootKit.kext
 sudo chown -R root:wheel build/com.YungRaj.MacRootKit.kext
